Extract stub tests into its own package
