+++ neo4j/test-integration/it_test.go
@@ -17,10 +17,10 @@
  * limitations under the License.
  */
 
-package integration_tests
+package test_integration
 
 import (
-	"github.com/neo4j/neo4j-go-driver/neo4j/integration-tests/control"
+	"github.com/neo4j/neo4j-go-driver/neo4j/test-integration/control"
 	. "github.com/onsi/ginkgo"
 )
