+++ neo4j/test-integration/transaction_test.go
@@ -17,14 +17,14 @@
  * limitations under the License.
  */
 
-package integration_tests
+package test_integration
 
 import (
 	"time"
 
 	"github.com/neo4j-drivers/gobolt"
 	. "github.com/neo4j/neo4j-go-driver/neo4j"
-	"github.com/neo4j/neo4j-go-driver/neo4j/integration-tests/control"
+	"github.com/neo4j/neo4j-go-driver/neo4j/test-integration/control"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 	"github.com/pkg/errors"
