+++ neo4j/test-integration/values_spatial_test.go
@@ -17,7 +17,7 @@
  * limitations under the License.
  */
 
-package integration_tests
+package test_integration
 
 import (
 	"math"
@@ -25,7 +25,7 @@ import (
 	"time"
 
 	. "github.com/neo4j/neo4j-go-driver/neo4j"
-	"github.com/neo4j/neo4j-go-driver/neo4j/integration-tests/control"
+	"github.com/neo4j/neo4j-go-driver/neo4j/test-integration/control"
 	. "github.com/neo4j/neo4j-go-driver/neo4j/internal/testing"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
