+++ neo4j/test-integration/values_temporal_test.go
@@ -17,14 +17,14 @@
  * limitations under the License.
  */
 
-package integration_tests
+package test_integration
 
 import (
 	"math/rand"
 	"time"
 
 	. "github.com/neo4j/neo4j-go-driver/neo4j"
-	"github.com/neo4j/neo4j-go-driver/neo4j/integration-tests/control"
+	"github.com/neo4j/neo4j-go-driver/neo4j/test-integration/control"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 )
