+++ neo4j/test/omega_error_matchers.go
@@ -17,7 +17,7 @@
  * limitations under the License.
  */
 
-package drivertest
+package test
 
 import (
 	"fmt"
